requirements.txt
flask
sqlalchemy
flask-jwt-extended
marshmallow-sqlalchemy
flask-sqlalchemy
flask-cors
pymysql
python-dotenv
requests
